# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "剪映小助手 API",
    "description": "视频编辑插件API，支持创建草稿、添加多媒体内容等功能",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://your-domain.com/api",
      "description": "Production server"
    }
  ],
  "paths": {
    "/create-draft": {
      "post": {
        "summary": "创建视频草稿",
        "description": "创建一个新的视频编辑草稿",
        "operationId": "createDraft",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "height": {
                    "type": "integer",
                    "description": "视频高度",
                    "minimum": 1
                  },
                  "width": {
                    "type": "integer",
                    "description": "视频宽度",
                    "minimum": 1
                  }
                },
                "required": ["height", "width"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "草稿创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "ids": { "type": "array", "items": { "type": "string" } }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/str-array": {
      "post": {
        "summary": "字符串数组转换",
        "description": "将输入字符串转换为数组",
        "operationId": "strArray",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "input": {
                    "type": "string",
                    "description": "输入字符串"
                  }
                },
                "required": ["input"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "转换成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "output": { "type": "array", "items": { "type": "string" } }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/audios-timelines": {
      "post": {
        "summary": "获取音频时间轴",
        "description": "根据音频ID获取对应的时间轴信息",
        "operationId": "audiosTimelines",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "音频ID数组"
                  }
                },
                "required": ["ids"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "timelines": { "type": "array", "items": { "type": "number" } }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/add-images": {
      "post": {
        "summary": "添加图片",
        "description": "向视频中添加图片元素",
        "operationId": "addImages",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "图片ID数组"
                  },
                  "images": {
                    "type": "array",
                    "items": { "type": "string", "format": "uri" },
                    "description": "图片URL数组"
                  },
                  "timelines": {
                    "type": "array",
                    "items": { "type": "number" },
                    "description": "时间轴数组"
                  },
                  "height": { "type": "number", "minimum": 0 },
                  "width": { "type": "number", "minimum": 0 },
                  "transform_x": { "type": "number" },
                  "transform_y": { "type": "number" },
                  "scale_x": { "type": "number", "minimum": 0 },
                  "scale_y": { "type": "number", "minimum": 0 },
                  "in_animation": { "type": "string" },
                  "in_animation_duration": { "type": "number", "minimum": 0 },
                  "out_animation": { "type": "string" },
                  "out_animation_duration": { "type": "number", "minimum": 0 }
                },
                "required": ["ids", "images", "timelines"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "图片添加成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/add-audios": {
      "post": {
        "summary": "添加音频",
        "description": "向视频中添加音频元素",
        "operationId": "addAudios",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "音频ID数组"
                  },
                  "mp3_uris": {
                    "type": "array",
                    "items": { "type": "string", "format": "uri" },
                    "description": "音频文件URL数组"
                  },
                  "timelines": {
                    "type": "array",
                    "items": { "type": "number" },
                    "description": "时间轴数组"
                  },
                  "audio_effect": { "type": "string" },
                  "volume": { "type": "number", "minimum": 0, "maximum": 1 }
                },
                "required": ["ids", "mp3_uris", "timelines"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "音频添加成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/add-captions": {
      "post": {
        "summary": "添加字幕",
        "description": "向视频中添加字幕",
        "operationId": "addCaptions",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "字幕ID数组"
                  },
                  "timelines": {
                    "type": "array",
                    "items": { "type": "number" },
                    "description": "时间轴数组"
                  },
                  "texts": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "字幕文本数组"
                  },
                  "text_color": { "type": "string" },
                  "font": { "type": "string" },
                  "font_size": { "type": "number", "minimum": 1 },
                  "border": { "type": "boolean" },
                  "border_color": { "type": "string" },
                  "transform_x": { "type": "number" },
                  "transform_y": { "type": "number" },
                  "alignment": { "type": "string" },
                  "line_spacing": { "type": "number" },
                  "letter_spacing": { "type": "number" },
                  "in_animation": { "type": "string" },
                  "in_animation_duration": { "type": "number", "minimum": 0 },
                  "out_animation": { "type": "string" },
                  "out_animation_duration": { "type": "number", "minimum": 0 }
                },
                "required": ["ids", "timelines", "texts"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "字幕添加成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/add-effects": {
      "post": {
        "summary": "添加特效",
        "description": "向视频中添加特效",
        "operationId": "addEffects",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "特效ID数组"
                  },
                  "timelines": {
                    "type": "array",
                    "items": { "type": "number" },
                    "description": "时间轴数组"
                  },
                  "effects": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "特效类型数组"
                  }
                },
                "required": ["ids", "timelines", "effects"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "特效添加成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/add-videos": {
      "post": {
        "summary": "添加视频",
        "description": "向项目中添加视频素材",
        "operationId": "addVideos",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "视频ID数组"
                  },
                  "timelines": {
                    "type": "array",
                    "items": { "type": "number" },
                    "description": "时间轴数组"
                  },
                  "video_uris": {
                    "type": "array",
                    "items": { "type": "string", "format": "uri" },
                    "description": "视频文件URL数组"
                  },
                  "volume": { "type": "number", "minimum": 0, "maximum": 1 },
                  "scale_x": { "type": "number", "minimum": 0 },
                  "scale_y": { "type": "number", "minimum": 0 },
                  "transform_x": { "type": "number" },
                  "transform_y": { "type": "number" }
                },
                "required": ["ids", "timelines", "video_uris"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "视频添加成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}